Break characteristic fetching out from initialize method, add passive scan method
Instead of fetching the characteristics automatically when initializing a BLE device object, this is done in a separate method.  That way the initialize method is fast and doesn't depend on spawning gatttool processes.  The passive scan method allows a scanner to scan for BLE devices without connecting to them (and thus keeping them from transmitting temporarily).