Add factory method to only create Usb objects from valid BluetoothLeDevice objects
Previously it was possible to try to initialize a Radbeacon Usb object from a BluetoothLeDevice that had some nil characteristics (possibly from a failed read).  This was causing exceptions in the utils methods when trying to convert those characteristics to human readable form.  The `create_if_valid` method is a factory method that will validate a BluetoothLeDevice before constructing a Usb object, that way you will never receive a Usb with missing attributes.